vulkan: the prefill tile reads its weights as words, not bytes
The driver issues one load per byte and joins none of them, 71 of them
in this kernel. The K-quant runs whose bytes are word aligned now take
one word per four values, and every scattered read extracts from the
word that holds it. qdecode.glsl reaches its bytes through a QBYTE
macro rather than naming an array, so a shader serves them from
whatever view it binds, and the tile kernel's byte binding goes away
rather than a word binding being added. Qwen3-8B-Q4_K_M prompt
processing rises a further 25 percent at 247 rows and Qwen3-0.6B-Q5_K_M
51 percent; the Q8_0 control moves both ways inside the layout band.

+++ src/backends/vulkan/shaders/q.glsl
@@ -40,8 +40,8 @@ float q8_0_scale(uint8_t bytes[2]) {
     return unpackHalf2x16(uint(bytes[0]) | (uint(bytes[1]) << 8)).x;
 }
 
-float half_at(uint8_t lo, uint8_t hi) {
-    return unpackHalf2x16(uint(lo) | (uint(hi) << 8)).x;
+float half_at(uint lo, uint hi) {
+    return unpackHalf2x16(lo | (hi << 8)).x;
 }
 
 // Bytes per block and values per block of a type, for the callers that
